fix(chatmessage): another futile attempt to fix the null.MeasureText runtime
https://github.com/ChaoticOnyx/OnyxBay/pull/13640

+++ code/datums/chatmessage.dm
@@ -149,7 +149,11 @@
 	var/static/regex/html_metachars = new(@"&[A-Za-z]{1,7};", "g")
 	var/complete_text = MAPTEXT("<span class='center[size ? " [size]" : ""]' style='color: [tgt_color]'>[text]</span>")
 
-	// Apparently, regexes work slow enough to let the client slip away before we reach this point. Luckily, everything below this check seems to be quick enought to not require even more checks. ~ToTh
+	// The weirdest things happen when we ASYNC.
+	if(QDELETED(src))
+		return
+
+	// Apparently, regexes work slow enough to let the client slip away before we reach this point. Luckily, everything below this check seems to be quick enough to not require even more checks. ~ToTh
 	// OR we can even get deleted by this point, nullifying 'owned_by'. I have no fucking idea.
 	if(QDELETED(owner) || !owner.client || !owned_by)
 		qdel(src)
